Se añadieron las clases e interfaces de mostrar los vuelos en addToFlight, se cargan las listas en el view desde el mainController

+++ src/Controllers/Refreshers/IRefreshMyFlights.java
@@ -11,6 +11,5 @@ import core.controllers.utils.Response;
  * @author HOLA
  */
 public interface IRefreshMyFlights {
-      public Response getMyFlightsRows(int passengerId);
-
+      public Response getMyFlightsRows(Long passengerId);
 }
